adjusted to 6 bytes for segmentid
I adjusted this so that we will have an 8 byte address, this will cleanly fit into a 4096 block 512 times. I thought about it more and 6 bytes for segmentid should be fine, that allows for 140 trillion segments. If we were writing 5 segments a second it would take 887,000+ years to run out of segmentid's. Even if storage is x1000 faster it would be good for 887 years.