Add Segment and Log classes (#8)
Log is an abstract representation of the log with methods for reading
and writing. When the filesystem is mounted an instance of this class
should be instantiated and used for all operations

I've represented segments as two different classes (with a common parent
class) so that ReadOnlySegment can store its data using an
immutable memoryview, which allows slicing without copying. Once
a ReadWriteSegment has been filled it should be converted into a
ReadOnlySegment.