Add MemoryCache and DiskCache for segments (#38)
Both of these inherit from the BackendWrapper abstract base class.
Children of this class this have all methods that they do not define
themselves forwarded to the backend they wrap. This allows us to create
a series of caching layers which are transparent to the caller and to
one another.

I expect the final system to look something like this:
Log -> MemoryCache -> DiskCache -> WriteQueue -> S3Bucket

MemoryCache keeps an in-memory cache of get_segment requests. DiskCache
tracks which segments are in its cache in memory, but stores the data
itself on disk. Both implement a least recently used eviction policy.

DiskCache must be instantiated in a "with" block to ensure that its
files are cleaned up when the filesystem is unmounted. To make this
possible I moved the responsibility of instantiating the bucket and
caches to mount.py, which then passes them to FuseApi.