Add AsyncWriter to make S3 writes asynchronous (#39)
This wraps a backend to handles writes asynchronously, using a thread
pool. It also maintains an in-memory cache of segments that are in the
process of being written. Otherwise a segment could be requested from
the backend before it has been written. Once each write completes the
segment is removed from this cache.