Fix bug with inode block address deserialization
Previously we were assuming that the data passed to INode.from_bytes()
would not be padded, and used its length to infer the length of each
BlockAddress. When the data was padded addresses other than the first
were not correctly deserialized.

Instead, this records the block address size as a constant, because it
is fixed. However, the actual block address deserialization logic does
not depend on this constant, so we should be careful to keep them in
sync.